ex4/requirements.txt
numpy 
Pillow
matplotlib
opencv-python